New test cases are added for validating the ListAlerts API with Zone filter
ListAlerts API has been enhanced to accept the zone to filter out the alerts.
For this new enhancement, new test cases has been added to validate the functionality
Test case 1: zone filter is included in the request and expecting the results
Test case2: zone filter is included in the request and expected ZERO results.